before keeping reformatting

+++ app/include/db_lmdb/db_lmdb_internal.h
@@ -10,11 +10,9 @@
 #ifndef DB_INTERNAL_H
 #define DB_INTERNAL_H
 
-#include "db_lmdb_config.h" /* */
-#include "db_lmdb_dbi.h"    /* dbi_desc_t */
+#include "db_lmdb.h"        /* interface dbi_decl_t */
+#include "db_lmdb_config.h" /* config file */
 #include "emlog.h"          /* EML_ERROR etc */
-
-#include <lmdb.h>
 
 #ifdef __cplusplus
 extern "C"
@@ -44,11 +42,11 @@ extern "C"
  */
 struct DB
 {
-    MDB_env*     env;                /* LMDB environment */
-    dbi_desc_t*  dbis;               /* array of DBI descriptors */
-    unsigned int n_dbis;             /* number of DBIs in array */
-    size_t       map_size_bytes;     /* current map size */
-    size_t       map_size_bytes_max; /* maximum map size */
+    MDB_env*    env;                /* LMDB environment */
+    dbi_desc_t* dbis;               /* array of DBI descriptors */
+    uint8_t     n_dbis;             /* number of DBIs in array */
+    size_t      map_size_bytes;     /* current map size */
+    size_t      map_size_bytes_max; /* maximum map size */
 
     // /* USER DBIs */
     // MDB_dbi db_user_id2meta; /* ID -> META */
